requirements.txt
numpy>=1.23.0
numpy<2.0.0 # sys_platform == 'darwin', # macOS OpenVINO errors https://github.com/ultralytics/ultralytics/pull/17221
matplotlib>=3.3.0
opencv-python>=4.6.0
pillow>=7.1.2
pyyaml>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.8.0
torch>=1.8.0,!=2.4.0 # sys_platform == 'win32', # Windows CPU errors w/ 2.4.0 https://github.com/ultralytics/ultralytics/issues/15049
torchvision>=0.9.0
tqdm>=4.64.0 # progress bars
psutil # system utilization
py-cpuinfo # display CPU info
pandas>=1.1.4
seaborn>=0.11.0 # plotting
ultralytics-thop>=2.0.0 # FLOPs computation https://github.com/ultralytics/thop